\name{Craft2003}
\alias{Craft2003}
\docType{data}
\title{18 studies of correlation coefficients reported by Craft et al. (2003)}
\description{
  This dataset includes 18 studies of correlation coefficients reported by Craft, Magyar, Becker, and Feltz in 2003.}
  \usage{data(Craft2003)}
\details{Using correlation coefficient as primary effect size
  measure, the main purposes of their meta-analysis were to examine the interrelationships
  between athletic performance and three subscales: cognitive anxiety, somatic anxiety, and self concept
  measured by the Competitive State Anxiety Inventory CSAI 2. For the
purpose of demonstration, I use a subset of the data that provides all six correlation
coefficients among cognitive anxiety, somatic anxiety, self concept, and sport
performance in athletes.
\tabular{ll}{
ID \tab ID for each study included\cr
N \tab sample size from each study included\cr
gender \tab gender\cr
p_male \tab percentage of male \cr
C1 \tab Correlation coefficient between cognitive anxiety and somatic anxiety \cr
C2 \tab Correlation coefficient between cognitive anxiety and self concept \cr
C3 \tab Correlation coefficient between cognitive anxiety and athletic performance \cr
C4 \tab Correlation coefficient between somatic anxiety and self concept \cr
C5 \tab Correlation coefficient between somatic anxiety and athletic performance \cr
C6 \tab Correlation coefficient between self concept and athletic performance \cr
}}
\examples{
  data(Craft2003)
}
\source{
\tabular{l}{
Ahn, S., Lu, M., Lefevor, G.T., Fedewa, A. and Celimli, S. (2016). Chapter 11. Application of meta-analysis in sport and exercise science. (pp. 233 253) In Ntoumanis, N. and Myers, N. (Eds.), An Introduction to Intermediate and Advanced Statistical Analyses for Sport and Exercise Scientists. John Wiley and Sons, Ltd.\cr
Cooper, Harris, Larry V. Hedges, and Jeffrey C. Valentine, eds. The handbook of research synthesis and meta analysis. Russell Sage Foundation, 2009.\cr}
}


